Extract color and decoration helpers from callback.rs (refs #88)
Pure-function extractions from crates/amux-render-gpu/src/callback.rs:

- color.rs — maybe_linearize, srgb_to_linear (sRGB conversion)
- decorations.rs — rasterize_curly_tile, rasterize_dotted_tile
  (underline decoration bitmaps)

These are free functions with no struct dependencies, so this is the
safest first step in decomposing the 1,583-line callback.rs module.
callback.rs drops to 1,468 lines. No behavior changes.
